requirements.txt
click==6.6
clize==3.0
diaspy-api==0.5.1
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.10.0
sigtools==2.0
six==1.10.0
Werkzeug==0.11.10
